updating fork

+++ runescape-api/src/main/java/net/runelite/rs/api/RSGameObject.java
@@ -63,7 +63,7 @@ public interface RSGameObject extends GameObject
 
 	@Import("hash")
 	@Override
-	int getHash();
+	long getHash();
 
 	@Import("flags")
 	int getFlags();
